+++ examples/scripts/compare_lead_acid_3D.py
@@ -46,7 +46,6 @@ param = models[0].default_parameter_values
 param.update(
     {
         "Current function [A]": 1,
-        "Bruggeman  coefficient": 0.001,
         "Initial State of Charge": 1,
         "Typical electrolyte concentration [mol.m-3]": 5600,
         "Negative electrode reference exchange-current density [A.m-2]": 0.08,