+++ pybamm/solvers/base_solver.py
@@ -233,7 +233,7 @@ class BaseSolver(object):
                 model.concatenated_rhs, model.concatenated_algebraic
             )
             # Process again, uses caching so should be quick
-            residuals, residuals_eval, jacobian_eval = process(all_states, "residuals",)
+            residuals, residuals_eval, jacobian_eval = process(all_states, "residuals")
             model.residuals_eval = residuals_eval
             model.jacobian_eval = jacobian_eval
             model.y0 = self.calculate_consistent_initial_conditions(model)
@@ -392,6 +392,10 @@ class BaseSolver(object):
         # Make sure model isn't empty
         if len(model.rhs) == 0 and len(model.algebraic) == 0:
             raise pybamm.ModelError("Cannot solve empty model")
+
+        # Make sure t_eval is monotonic
+        if (np.diff(t_eval) < 0).any():
+            raise pybamm.SolverError("t_eval must increase monotonically")
 
         # Set up
         timer = pybamm.Timer()
@@ -589,14 +593,18 @@ class SolverCallable:
 
     def __call__(self, t, y):
         y = y[:, np.newaxis]
-        if self.name in ["RHS", "algebraic", "residuals"]:
-            return self.function(t, y)[:, 0]
+        if self.name in ["RHS", "algebraic", "residuals", "event"]:
+            return self.function(t, y).flatten()
         else:
             return self.function(t, y)
 
     def function(self, t, y):
         if self.form == "casadi":
-            return self._function(t, y, self.inputs).full()
+            if self.name in ["RHS", "algebraic", "residuals", "event"]:
+                return self._function(t, y, self.inputs).full()
+            else:
+                # keep jacobians sparse
+                return self._function(t, y, self.inputs)
         else:
             return self._function(t, y, self.inputs, known_evals={})[0]