+++ tests/unit/test_parameters/test_parameter_values.py
@@ -20,6 +20,10 @@ class TestParameterValues(unittest.TestCase):
         # from dict
         param = pybamm.ParameterValues({"a": 1})
         self.assertEqual(param["a"], 1)
+        self.assertEqual(list(param.keys())[0], "a")
+        self.assertEqual(list(param.values())[0], 1)
+        self.assertEqual(list(param.items())[0], ("a", 1))
+
         # from file
         param = pybamm.ParameterValues(
             values="input/parameters/lithium-ion/cathodes/lico2_Marquis2019/"
@@ -52,11 +56,16 @@ class TestParameterValues(unittest.TestCase):
         self.assertEqual(param["a"], 2)
         # with conflict
         param.update({"a": 3})
-        self.assertEqual(param["a"], 3)
+        # via __setitem__
+        param["a"] = 2
+        self.assertEqual(param["a"], 2)
         with self.assertRaisesRegex(
-            ValueError, "parameter 'a' already defined with value '3'"
+            ValueError, "parameter 'a' already defined with value '2'"
         ):
             param.update({"a": 4}, check_conflict=True)
+        # with parameter not existing yet
+        with self.assertRaisesRegex(KeyError, "Cannot update parameter"):
+            param.update({"b": 1})
 
     def test_check_and_update_parameter_values(self):
         # Can't provide a current density of 0, as this will cause a ZeroDivision error
@@ -367,6 +376,7 @@ class TestParameterValues(unittest.TestCase):
                 "cathodes",
                 "lico2_Marquis2019",
             ),
+            check_already_exists=False,
         )
 
         a = pybamm.Parameter("a")